LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-sends-reports-on-violation-2.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Report is observable to ReportingObserver assert_equals: expected (string) "https://web-platform.test:9443/content-security-policy/reporting-api/reporting-api-sends-reports-on-violation-2.https.sub.html" but got (object) null
PASS Violation report status OK.
